requirements.txt
absl-py==1.4.0
aiohttp==3.8.1
aiohttp-cors==0.7.0
aiosignal==1.3.1
alabaster==0.7.13
anyio==3.7.1
apache-beam==2.46.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
backports.functools-lru-cache==1.6.5
beautifulsoup4==4.12.2
bleach==6.0.0
blessed==1.20.0
blinker==1.6.2
brotlipy==0.7.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.0
chardet==5.2.0
charset-normalizer==2.0.0
cirq==1.2.0
cirq-aqt==1.2.0
cirq-core==1.2.0
cirq-ft==1.2.0
cirq-google==1.2.0
cirq-ionq==1.2.0
cirq-pasqal==1.2.0
cirq-rigetti==1.2.0
cirq-web==1.2.0
click==8.1.6
cloudpickle==2.2.1
colorful==0.5.5
comm==0.1.4
contextlib2==21.6.0
contourpy==1.1.0
crcmod==1.7
cryptography==41.0.2
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.1.1
distlib==0.3.7
dm-reverb==0.11.0
dm-tree==0.1.8
docopt==0.6.2
docutils==0.18.1
duet==0.2.8
entrypoints==0.4
exceptiongroup==1.1.2
executing==1.2.0
fastavro==1.8.2
fasteners==0.18
fastjsonschema==2.18.0
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.41.1
fqdn==1.5.1
frozenlist==1.3.3
gast==0.4.0
gin-config==0.5.0
google-api-core==2.11.1
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
google-vizier==0.1.6
googleapis-common-protos==1.59.1
gpustat==1.1
greenlet==2.0.2
grpcio==1.48.2
grpcio-status==1.56.2
grpcio-tools==1.48.2
gym==0.23.0
gym-notices==0.0.8
h11==0.14.0
h5py==3.9.0
hdfs==2.7.0
httpcore==0.16.3
httplib2==0.21.0
httpx==0.23.3
idna==3.4
imagesize==1.4.1
importlib-metadata==6.8.0
importlib-resources==6.0.0
ipykernel==6.15.0
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==7.8.0
iso8601==1.1.0
isoduration==20.11.0
jax==0.4.13
jedi==0.18.2
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_server==2.7.2
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab_server==2.24.0
jupyterlab-widgets==1.1.5
keras==2.12.0
kiwisolver==1.4.4
lark==0.11.3
libclang==16.0.6
line-profiler==4.1.1
llvmlite==0.40.1
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mistune==3.0.1
ml-collections==0.1.1
ml-dtypes==0.2.0
more-itertools==10.0.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.2
nbclient==0.8.0
nbconvert==7.7.2
nbformat==5.9.1
nbsphinx==0.9.2
nest-asyncio==1.5.6
networkx==2.8.8
notebook==7.0.2
notebook_shim==0.2.3
ntlm-auth==1.5.0
numba==0.57.1
numpy==1.22.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cudnn-cu11==8.6.0.163
nvidia-ml-py==12.535.77
oauthlib==3.2.2
objsize==0.6.1
opencensus==0.11.2
opencensus-context==0.1.3
opt-einsum==3.3.0
orjson==3.9.2
overrides==7.4.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pbr==5.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pip==23.1.2
platformdirs==3.9.1
ply==3.11
portpicker==1.5.2
prometheus-client==0.17.1
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py-spy==0.3.14
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.21
pydantic==1.10.12
pydot==1.4.2
pygame==2.1.3
Pygments==2.15.1
PyJWT==2.8.0
pymongo==3.13.0
pyOpenSSL==23.2.0
pyparsing==3.0.9
pyperclip==1.8.2
pyquil==3.2.1
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
python-rapidjson==1.10
pytz==2023.3
pyu2f==0.1.5
PyYAML==6.0.1
pyzmq==25.1.1
pyzx==0.7.3
qcs-api-client==0.21.5
qiskit==0.43.3
qiskit-aer==0.12.2
qiskit-ibmq-provider==0.20.2
qiskit-terra==0.24.2
qutip==4.7.2
ray==2.6.2
referencing==0.30.2
regex==2023.6.3
requests==2.31.0
requests-ntlm==1.1.0
requests-oauthlib==1.3.1
retry==0.9.2
retrying==1.3.4
rfc3339==6.2
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rlds==0.1.8
rpcq==3.11.0
rpds-py==0.9.2
rsa==4.9
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
rustworkx==0.13.0
scipy==1.11.1
Send2Trash==1.8.2
setuptools==68.0.0
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.4.1
Sphinx==6.2.1
sphinx-rtd-theme==1.2.2
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.0
stack-data==0.6.2
stevedore==5.1.0
stim==1.11.0
symengine==0.9.2
sympy==1.12
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-gnn==0.5.1
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-probability==0.20.1
termcolor==2.3.0
terminado==0.17.1
tf-agents==0.17.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tornado==6.3.3
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.15
virtualenv==20.21.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
websockets==11.0.3
Werkzeug==2.3.5
wheel==0.38.4
widgetsnbextension==3.6.5
wrapt==1.14.1
yarl==1.7.2
zipp==3.16.2
zstandard==0.21.0
